% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIJ1.R
\name{getFloatMatrix}
\alias{getFloatMatrix}
\title{Get float values from the ImageProcessor!}
\usage{
getFloatMatrix(ip)
}
\arguments{
\item{ip}{The ip argument.}
}
\description{
Get float values from the ImageProcessor!
}
\examples{
floatMatrix<-getFloatMatrix(ip)
}
