% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIJ1.R
\name{getImageProcessor}
\alias{getImageProcessor}
\title{Get a reference to the ImageProcessor class!}
\usage{
getImageProcessor(im)
}
\arguments{
\item{im}{The im argument.}
}
\description{
Get a reference to the ImageProcessor class!
}
\examples{
ip<-getImageProcessor(imagePlus)
}
