% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIJ1.R
\name{getIntegerMatrix}
\alias{getIntegerMatrix}
\title{Get integer values from the ImageProcessor!}
\usage{
getIntegerMatrix(ip)
}
\arguments{
\item{ip}{The ip argument.}
}
\description{
Get integer values from the ImageProcessor!
}
\examples{
integerMatrix<-getIntegerMatrix(ip)
}
