% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIJ1.R
\name{loadImage}
\alias{loadImage}
\title{Open an image!}
\usage{
loadImage(path)
}
\arguments{
\item{path}{The path argument.}
}
\description{
Open an image!
}
\examples{
imagePlus<-loadImage('http://imagej.nih.gov/ij/images/blobs.gif')
}
