% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIJ1.R
\name{plotIJ1Coords}
\alias{plotIJ1Coords}
\title{Transpose the values for the correct image matrix! intValues<-t(intValues)
Plot the integer values with origin in the upper left!}
\usage{
plotIJ1Coords(intValues)
}
\arguments{
\item{intValues}{The values argument.}
}
\description{
Transpose the values for the correct image matrix! intValues<-t(intValues)
Plot the integer values with origin in the upper left!
}
\examples{
plotIJ1Coords(floatMatrix)
}
